update readme links (#3)
This pull request includes changes to the `README.md` file to correct
markdown link syntax and update the OpenAPI documentation link.

Markdown link corrections:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L10-R10):
Corrected the syntax for the "Ecoland-Apps" link to properly format the
markdown.
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L19-R20):
Corrected the syntax for the "Docs" and "OpenAPI" links to properly
format the markdown and updated the OpenAPI link to a relative path.