add-more-test (#9)
This pull request includes several changes focused on adding tests and
improving code quality in the `go` package. The most important changes
include the addition of new test cases for various functionalities and
the removal of a debug log statement.

### Addition of new test cases:

*
[`go/authentication/authentication_test.go`](diffhunk://#diff-786a9475fa9f1da8c321abfae9636ae00428c34415912e03fa013c9e210135e1R1-R47):
Added tests for `CreateNewJWT` and `ValidateJWT` functions to ensure JWT
creation and validation work correctly.
*
[`go/database/databse_test.go`](diffhunk://#diff-af603464515f45d42a94f5f28b46cc79b089dd729dc91f03e74bf65e05ffc644R1-R24):
Added tests for `SetDB` and `GetDB` functions to verify database
connection handling.
*
[`go/utils/header_test.go`](diffhunk://#diff-e46f8e26ebf60155940133c2374ddc368fe4350876a40d9577f8f09e4732f800R1-R16):
Added a test for `SetHeaderJson` function to ensure the correct content
type is set in the response header.
*
[`go/utils/reqMethod_test.go`](diffhunk://#diff-40c3ca065d582a900ac958a8432f5e756e6fd8c9942b0db609267d94a9afb31aR1-R33):
Added a test for `IsMethodPOST` function to check if the HTTP request
method is correctly identified as POST.

### Code quality improvement:

*
[`go/authentication/authentication.go`](diffhunk://#diff-25b3f8e85e167774bc693abd12693a6d35f4876855be6c2cf61c11f414d44220L46-L52):
Removed a debug log statement from the `ValidateJWT` function to clean
up the code.