check if email exists (#19)
This pull request introduces several changes to the user management
functionality in the codebase. The primary changes include adding an
email field to the `users` table, implementing a function to check if an
email already exists, and updating the signup process to validate email
addresses.

Changes to user management:

*
[`go/database/databaseTypes.go`](diffhunk://#diff-ba1f3e49d3895fd272ccd00f83587663c1fc7ed59e9a9b97995d7a5c5cfed1a8R75):
Added `UserTable` constant for the `users` table.
*
[`go/database/dbUser.go`](diffhunk://#diff-72d7af4d064cc8cbeb4f1b903796f4a142b1901cf5a5c3ec85140264f6fc7470R51-R60):
Introduced `EmailExists` function to check if an email already exists in
the `users` table.
*
[`go/service/user.go`](diffhunk://#diff-bb507dd6db15ed3c8c6cbdefb91e2d8655227fb2b1e4ee3a905222f530ee5a9aR119-R127):
Updated the `SignUp` function to include email validation and check for
existing emails before proceeding with user registration.
*
[`go/service/user.go`](diffhunk://#diff-bb507dd6db15ed3c8c6cbdefb91e2d8655227fb2b1e4ee3a905222f530ee5a9aL188-R202):
Modified the `AddUser` function to include the email field when
inserting a new user into the `users` table.

Initialization improvements:

*
[`go/main.go`](diffhunk://#diff-a240b23c8c4a5b71406f4393240035163be74a3840d5fdaa4b2a993872e8cf65R42):
Added a call to `InitDatabaseTables` during the initial setup to ensure
database tables are initialized.